Reorganize for logical package structure per service
