--- libs/langgraph-checkpoint-aws/langgraph_checkpoint_aws/agentcore/valkey/__init__.py
@@ -1,13 +1,0 @@
-"""
-AgentCore Valkey integration for LangGraph checkpoint saving.
-
-This module provides AgentCore-compatible checkpoint savers that use Valkey
-as the storage backend, combining AgentCore session management concepts
-with Valkey's high-performance storage and search capabilities.
-"""
-
-from .saver import AgentCoreValkeySaver
-
-__all__ = [
-    "AgentCoreValkeySaver",
-]
