+++ libs/langgraph-checkpoint-aws/langgraph_checkpoint_aws/checkpoint/valkey/agentcore/models.py
@@ -8,7 +8,7 @@ from typing import Any
 
 from pydantic import BaseModel, ConfigDict, Field
 
-from ..models import CheckpointerConfig
+from ...agentcore.models import CheckpointerConfig
 
 
 class ValkeyCheckpointerConfig(CheckpointerConfig):
