+++ libs/langgraph-checkpoint-aws/langgraph_checkpoint_aws/store/agentcore/__init__.py
@@ -1,0 +1,7 @@
+"""AgentCore store implementations."""
+
+from .store import AgentCoreMemoryStore
+
+__all__ = [
+    "AgentCoreMemoryStore",
+]
