+++ libs/langgraph-checkpoint-aws/tests/unit_tests/store/dynamodb/test_dynamodb_store.py
@@ -26,7 +26,7 @@ from langgraph_checkpoint_aws.store.dynamodb.exceptions import (
 # Test constants for async testing
 N_ASYNC_CALLS = 5
 MOCK_SLEEP_DURATION = 0.1 / N_ASYNC_CALLS
-OVERHEAD_DURATION = 0.01
+OVERHEAD_DURATION = 0.03
 TOTAL_EXPECTED_TIME = MOCK_SLEEP_DURATION + OVERHEAD_DURATION
